fix(suite-native): fix address formatter
-  It was not working for account labels because it was using the address as the accountKey.
- The last change caused it to crash.
-  Removing the old, non-functional functionality. AddressFormatter can be merged with AddressLabel as a follow-up.
