+++ suite-native/formatters/package.json
@@ -14,7 +14,6 @@
         "@mobily/ts-belt": "^3.13.1",
         "@reduxjs/toolkit": "2.11.0",
         "@suite-common/formatters": "workspace:*",
-        "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
