--- suite-native/formatters/src/components/AccountAddressFormatter.tsx
@@ -1,51 +1,0 @@
-import { Platform } from 'react-native';
-import { useSelector } from 'react-redux';
-
-import { SuiteSyncDataRootState, selectSuiteSyncAccountLabel } from '@suite-common/suite-sync';
-import { AccountsRootState } from '@suite-common/wallet-core';
-import { AccountKey } from '@suite-common/wallet-types';
-import { parseAccountKey, parseDeviceStaticSessionId } from '@suite-common/wallet-utils';
-import { Text, TextProps } from '@suite-native/atoms';
-import { mergeNativeStyleObjects, prepareNativeStyle, useNativeStyles } from '@trezor/styles';
-
-import { FormatterProps } from '../types';
-
-type AccountAddressFormatterProps = FormatterProps<AccountKey> & TextProps;
-
-const addressStyle = prepareNativeStyle(_ => ({
-    // ellipsizeMode="middle" is not working on Android with negative letterSpacing defined in @trezor/theme typography.
-    extend: {
-        condition: Platform.OS === 'android',
-        style: {
-            letterSpacing: 0,
-        },
-    },
-}));
-
-export const AccountAddressFormatter = ({
-    value,
-    style,
-    ...rest
-}: AccountAddressFormatterProps) => {
-    const { applyStyle } = useNativeStyles();
-
-    const { accountDescriptor, networkSymbol, deviceStaticSessionId } = parseAccountKey(
-        value as AccountKey, // Todo: this is WTF, there is something fishy with this component see `FormatterProps<AccountKey>`
-    );
-    const { walletDescriptor } = parseDeviceStaticSessionId(deviceStaticSessionId);
-
-    const accountLabel = useSelector((state: AccountsRootState & SuiteSyncDataRootState) =>
-        selectSuiteSyncAccountLabel(state, walletDescriptor, accountDescriptor, networkSymbol),
-    );
-
-    const baseAddressStyle = applyStyle(addressStyle);
-    const mergedAddressStyle = style
-        ? mergeNativeStyleObjects([style, baseAddressStyle])
-        : baseAddressStyle;
-
-    return (
-        <Text numberOfLines={1} ellipsizeMode="middle" style={mergedAddressStyle} {...rest}>
-            {accountLabel || value}
-        </Text>
-    );
-};
